feat: P0-P2 release engineering completion
P0 (Release Engineering):
- fix: dead_code warning in code-sandbox-plugin (use self.status)
- feat: start.sh --china flag (China network optimization)
- feat: start.sh --with-openclaw flag (OpenClaw Bridge integration)
- feat: openclaw-bridge HTTP mode support (HTTP_PORT env)
- fix: CI clippy config (meaningful lint levels)
- docs: README updated (44 crates, new flags, WeChat config, OpenClaw)

P1 (Ecosystem):
- feat: start.sh update enhanced (cargo install, version changelog)
- feat: WebUI auto-build in start.sh (trunk + wasm detection)
- fix: SQLite migration 005 registered (was missing)

P2 (Extensions):
- feat: web-search plugin (DuckDuckGo search engine)
- feat: Tauri desktop scaffolding (cross-platform desktop app)
- docs: desktop build instructions in README

+++ backends/src/llm_llmkit.rs
@@ -113,7 +113,7 @@ impl Llm for LlmkitLlm {
             req = req.with_system(sys);
         }
         if let Some(max_tokens) = request.max_tokens {
-            req = req.with_max_tokens(max_tokens as u32);
+            req = req.with_max_tokens(max_tokens);
         }
         if let Some(temp) = request.temperature {
             req = req.with_temperature(temp as f32);
